requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.3.3
backcall==0.1.0
billiard==3.6.3.0
decorator==4.4.2
Django==3.0.4
django-crontab==0.7.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-widget-tweaks==1.4.8
importlib-metadata==1.6.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
typed-ast==1.4.1
vine==1.3.0
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.1.0
